Cleanup compiler warnings. Remove unused vars. Change g15r_drawXBM declaration to non-static so it can be used by apps.
git-svn-id: https://g15tools.svn.sourceforge.net/svnroot/g15tools/trunk@300 f3f12dcf-2114-0410-a68b-d6aaca122e96

+++ libg15render/src/g15fontconvert.c
@@ -20,6 +20,7 @@
 #include "libg15render.h"
 #include <sys/types.h>
 #include <fcntl.h>
+#include <unistd.h>
 #ifdef TTF_SUPPORT
 #include <ft2build.h>
 #include FT_FREETYPE_H
@@ -82,11 +83,11 @@ int convertG15Font(char *inFilename, char *oFilename, int size, int gap){
     if(font->ascender_height==0) {
         font->ascender_height = font->font_height;
         face->size->metrics.ascender = font->font_height<<6;
-        printf("modded ascender to %i\n", face->size->metrics.ascender>>6);
+        printf("modded ascender to %i\n", (int)face->size->metrics.ascender>>6);
     }
     if(face->size->metrics.descender==0) {
         face->size->metrics.descender = (font->font_height<<6) - (face->available_sizes->size)  ;
-        printf("modded descender to %i\n",face->size->metrics.descender>>6);
+        printf("modded descender to %i\n",(int)face->size->metrics.descender>>6);
          font->font_height = (face->size->metrics.ascender >> 6) + (face->size->metrics.descender >> 6);
     }
     if(font->lineheight==0) {
